[Interactive Graph Editor] Move the 'Add locked figure' and 'Collapse/Expand all' buttons to the top. Put the locked figures section under a collapsible
