+++ .changeset/thirty-trees-work.md
@@ -1,0 +1,5 @@
+---
+"@khanacademy/perseus-editor": minor
+---
+
+[Interactive Graph Editor] Move the "Add locked figure" and "Collapse/Expand all" buttons to the top of the locked figures section. Put the locked figures section under its own collapsible section.
