+++ packages/perseus-editor/src/components/locked-figure-select.tsx
@@ -5,7 +5,6 @@
  *
  * Used in the interactive graph editor's locked figures section.
  */
-import {View} from "@khanacademy/wonder-blocks-core";
 import {ActionItem, ActionMenu} from "@khanacademy/wonder-blocks-dropdown";
 import {spacing, color} from "@khanacademy/wonder-blocks-tokens";
 import {StyleSheet} from "aphrodite";
@@ -20,39 +19,35 @@ const LockedFigureSelect = (props: Props) => {
     const {id, onChange} = props;
 
     return (
-        <View style={styles.container}>
-            <ActionMenu
-                menuText="Add locked figure"
-                style={styles.addElementSelect}
-            >
-                {[
-                    <ActionItem
-                        key={`${id}-point`}
-                        label="Point"
-                        onClick={() => onChange("point")}
-                    >
-                        Point
-                    </ActionItem>,
-                    <ActionItem
-                        key={`${id}-line`}
-                        label="Line"
-                        onClick={() => onChange("line")}
-                    >
-                        Line
-                    </ActionItem>,
-                ]}
-            </ActionMenu>
-        </View>
+        <ActionMenu
+            menuText="Add locked figure"
+            style={styles.addElementSelect}
+        >
+            {[
+                <ActionItem
+                    key={`${id}-point`}
+                    label="Point"
+                    onClick={() => onChange("point")}
+                >
+                    Point
+                </ActionItem>,
+                <ActionItem
+                    key={`${id}-line`}
+                    label="Line"
+                    onClick={() => onChange("line")}
+                >
+                    Line
+                </ActionItem>,
+            ]}
+        </ActionMenu>
     );
 };
 
 const styles = StyleSheet.create({
-    container: {
-        marginTop: spacing.xSmall_8,
-    },
     addElementSelect: {
         backgroundColor: color.fadedBlue8,
         borderRadius: spacing.xxxSmall_4,
+        height: spacing.xLarge_32,
     },
 });
